+++ internal/adapters/api/rest/api.go
@@ -5,13 +5,15 @@ import (
 	"fmt"
 	"net"
 	"net/http"
+
+	"github.com/k0st1a/gophermart/internal/ports"
 )
 
 type api struct {
 	server *http.Server
 }
 
-func NewAPI(ctx context.Context, address string, storage any) *api {
+func NewAPI(ctx context.Context, address string, storage ports.UserStorage) *api {
 	h := newHandler(storage)
 
 	r := newRouter()
