+++ internal/adapters/db/storage.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/rs/zerolog/log"
 )
 
 type db struct {
@@ -25,6 +26,18 @@ func NewDB(ctx context.Context, dsn string) (*db, error) {
 	return &db{
 		pool: pool,
 	}, nil
+}
+
+func (d *db) CreateUser(ctx context.Context, login, password string) (int64, error) {
+	log.Printf("RegisterUser, login:%s, password:%s", login, password)
+	var id int64
+
+	err := d.pool.QueryRow(ctx, "INSERT INTO users (login,password) VALUES($1,$2) RETURNING id", login, password).Scan(&id)
+	if err != nil {
+		return id, fmt.Errorf("failed to create user:%w", err)
+	}
+
+	return id, nil
 }
 
 func (d *db) Close() {
